Refactor doctor-config-flow.ts into smaller modules
