+++ src/commands/doctor-config/utils.ts
@@ -1,0 +1,99 @@
+import type { ZodIssue } from "zod";
+import type { OpenClawConfig } from "../../config/config.js";
+import { OpenClawSchema } from "../../config/config.js";
+
+export type UnrecognizedKeysIssue = ZodIssue & {
+  code: "unrecognized_keys";
+  keys: PropertyKey[];
+};
+
+export function normalizeIssuePath(path: PropertyKey[]): Array<string | number> {
+  return path.filter((part): part is string | number => typeof part !== "symbol");
+}
+
+export function isUnrecognizedKeysIssue(issue: ZodIssue): issue is UnrecognizedKeysIssue {
+  return issue.code === "unrecognized_keys";
+}
+
+export function formatPath(parts: Array<string | number>): string {
+  if (parts.length === 0) {
+    return "<root>";
+  }
+  let out = "";
+  for (const part of parts) {
+    if (typeof part === "number") {
+      out += `[${part}]`;
+      continue;
+    }
+    out = out ? `${out}.${part}` : part;
+  }
+  return out || "<root>";
+}
+
+export function resolvePathTarget(root: unknown, path: Array<string | number>): unknown {
+  let current: unknown = root;
+  for (const part of path) {
+    if (typeof part === "number") {
+      if (!Array.isArray(current)) {
+        return null;
+      }
+      if (part < 0 || part >= current.length) {
+        return null;
+      }
+      current = current[part];
+      continue;
+    }
+    if (!current || typeof current !== "object" || Array.isArray(current)) {
+      return null;
+    }
+    const record = current as Record<string, unknown>;
+    if (!(part in record)) {
+      return null;
+    }
+    current = record[part];
+  }
+  return current;
+}
+
+export function stripUnknownConfigKeys(config: OpenClawConfig): {
+  config: OpenClawConfig;
+  removed: string[];
+} {
+  const parsed = OpenClawSchema.safeParse(config);
+  if (parsed.success) {
+    return { config, removed: [] };
+  }
+
+  const next = structuredClone(config);
+  const removed: string[] = [];
+  for (const issue of parsed.error.issues) {
+    if (!isUnrecognizedKeysIssue(issue)) {
+      continue;
+    }
+    const path = normalizeIssuePath(issue.path);
+    const target = resolvePathTarget(next, path);
+    if (!target || typeof target !== "object" || Array.isArray(target)) {
+      continue;
+    }
+    const record = target as Record<string, unknown>;
+    for (const key of issue.keys) {
+      if (typeof key !== "string") {
+        continue;
+      }
+      if (!(key in record)) {
+        continue;
+      }
+      delete record[key];
+      removed.push(formatPath([...path, key]));
+    }
+  }
+
+  return { config: next, removed };
+}
+
+export function asObjectRecord(value: unknown): Record<string, unknown> | null {
+  if (!value || typeof value !== "object" || Array.isArray(value)) {
+    return null;
+  }
+  return value as Record<string, unknown>;
+}
